fixed setup.sh to not relink folders

+++ setup.sh
@@ -4,7 +4,9 @@ echo Linking files to home directory
 
 for file in *.sym
 do
-   ln -s `pwd`/$file ~/.${file%.sym}
+  if [ -f $file ]; then
+    ln -s `pwd`/$file ~/.${file%.sym}
+  fi
 done
 
 # copy other config files
